+++ functions/package.json
@@ -34,5 +34,9 @@
   },
   "peerDependencies": {
     "typescript": "^4.0.0"
+  },
+  "dependencies": {
+    "axios": "^0.21.1",
+    "middy": "^0.36.0"
   }
 }
