+++ functions/src/hello/index.ts
@@ -1,17 +1,19 @@
-import { APIGatewayEvent, APIGatewayProxyResult } from "aws-lambda";
+import { APIGatewayEvent, APIGatewayProxyResult, APIGatewayProxyHandler } from "aws-lambda";
+import middy from "middy"
+import { myFXBookLoginMiddleware } from "../middleware/MyFXBookLogin"
 
 export interface HelloResponse {
   message: string;
 }
 
-export const handler = async (
+const helloHandler = async (
   event: APIGatewayEvent
 ): Promise<APIGatewayProxyResult> => {
   // your server-side functionality
 
   const msg = event.queryStringParameters?.msg;
   const response: HelloResponse = { message: "Hello World " + msg };
-
+  console.log(event.headers);
   return {
     statusCode: 200,
     headers: {
@@ -20,3 +22,5 @@ export const handler = async (
     body: JSON.stringify(response),
   };
 };
+
+export const handler: APIGatewayProxyHandler = middy(helloHandler).use(myFXBookLoginMiddleware())
